Fix #349 retry on jobInternalError (#363)

+++ kcbq-connector/src/main/java/com/wepay/kafka/connect/bigquery/utils/SleepUtils.java
@@ -4,7 +4,7 @@ import java.util.concurrent.ThreadLocalRandom;
 
 public final class SleepUtils {
 
-    public static void waitRandomTime(int sleepMs, int jitterMs) throws InterruptedException {
-        Thread.sleep(sleepMs + ThreadLocalRandom.current().nextInt(jitterMs));
+    public static void waitRandomTime(long sleepMs, long jitterMs) throws InterruptedException {
+        Thread.sleep(sleepMs + ThreadLocalRandom.current().nextLong(jitterMs));
     }
 }
